backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
click>=7.1,<8.2
fastapi==0.119.1
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.11
iniconfig==2.0.0
joblib==1.5.2
nltk==3.9.2
packaging==25.0
pluggy==1.6.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pytest==8.4.2
python-multipart==0.0.20
regex==2025.10.23
sniffio==1.3.1
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
gTTS==2.5.3
supabase==2.10.0
python-dotenv==1.0.0
